+++ integration_test/tests/index.ts
@@ -3,6 +3,7 @@ import audioTrackTest from './audioTrackTest';
 import CaptionTest from './captionTest';
 import ErrorTest from './errorTest';
 import LoadingTest from './loadingTest';
+import MediaControlsTest from './mediaControlsTest';
 import metadataId3Test from './metadataId3Test';
 import PlaybackTest from './playbackTest';
 import UnloadingTest from './unloadingTest';
@@ -13,6 +14,7 @@ export default [
   CaptionTest,
   ErrorTest,
   LoadingTest,
+  MediaControlsTest,
   metadataId3Test,
   PlaybackTest,
   UnloadingTest,